update zmq loadbalancer to reuse the same worker most of the time
This is mostly like https://github.com/CanalTP/navitia/pull/2551
This change aims to reduce the number of allocation done by kraken if it
doesn't have to do it.
Currently available workers are stored in a
queue, so every workers will handle a request a some point in time no
matter how much concurrent request there is.

Workers are now stored in a stack, so the last worker used will be
reused for handle the next requests. This will reduce the cost of
initializing workers when there isn't that many concurrent requests.

This is a optimization for big coverage (think transilien) that have a
lot of thread (80) to handle peak, but most of the time (outside of load
testing...) there are rarely more than a few concurrents requests on
kraken (4).

This should reduce response time after a data update or realtime update
and reduce memory consumption.